chore: separate code responsibilities
