+++ src/app/paintings/components/PaintingCard.tsx
@@ -9,6 +9,7 @@ import { motion } from "framer-motion";
 import type { Painting } from "@/types/types";
 
 import { Button } from "@/components/ui/button";
+import { usePaintingAnimations } from "./usePaintingAnimations";
 
 const MotionImage = motion.create(Image);
 
@@ -23,53 +24,15 @@ export const PaintingCard = memo(function PaintingCard({
   onCardClick: (id: string) => void;
   index: number;
 }) {
-  const imageVariants = {
-    initial: { y: 0, scale: 1.015 },
-    hover: {
-      y: 30,
-      scale: 1.25,
-      transition: { duration: 0.5, ease: "easeInOut" },
-    },
-  } as const;
-
-  const overlayVariants = {
-    initial: { opacity: 0, scale: 1 },
-    hover: {
-      opacity: 1,
-      scale: 1.25,
-      y: 30,
-      transition: { duration: 0.5, ease: "easeInOut" },
-    },
-  } as const;
-
-  const titleVariants = {
-    initial: { x: 300 },
-    hover: { x: 0, transition: { duration: 0.3, ease: "easeInOut" } },
-  } as const;
-
-  const dateVariants = {
-    initial: { x: -300 },
-    hover: { x: 0, transition: { duration: 0.3, ease: "easeInOut" } },
-  } as const;
-
-  const descriptionTitleVariants = {
-    initial: { x: 0 },
-    hover: { x: -400, transition: { duration: 0.3, ease: "easeInOut" } },
-  } as const;
-
-  const descriptionDateVariants = {
-    initial: { x: 0 },
-    hover: { x: 400, transition: { duration: 0.3, ease: "easeInOut" } },
-  } as const;
-
-  const buttonVariants = {
-    initial: { opacity: 1, pointerEvents: "none" },
-    hover: {
-      opacity: 1,
-      pointerEvents: "auto",
-      transition: { duration: 0.3, ease: "easeInOut" },
-    },
-  } as const;
+  const {
+    imageVariants,
+    overlayVariants,
+    titleVariants,
+    dateVariants,
+    buttonVariants,
+    descriptionTitleVariants,
+    descriptionDateVariants,
+  } = usePaintingAnimations();
 
   const handleKeyDown = (event: KeyboardEvent<HTMLElement>) => {
     if (event.key === "Enter" || event.key === " ") {
